Move getShapeName out of types/style
